Augmenting `fz_sanitize_path_ex()` with a `size_t start_at_offset` argument: When `start_at_offset` is not zero (i.e. start-of-string), only the part of the path at & following the `start_at_offset` position is sanitized. It is assumed that the non-zero offset ALWAYS points past any critical UNC or MSWindows Drive designation parts of the path.
The reason for this is that we can now apply this path sanitization to any bulktest output-remapped file path and/or other spots where we only wish to 'clean/strip' the trailing (and probably *generated*) part of a file path, instead of going over the entire thing... but, you ask, it wouldn't harm to let the entire path be processed for it would be safe, surely? Well, when we want a slightly more restrictive/pedantic cleanup, we DO NOT want that restrictiveness applied to any overarching basee directory path or some such, only to the part of the path that we appended to that base segment. This is where the offset argument comes in handy.

- fix added: ':' colon is now in the default illegal-for-directory-path-part character set: while it is legal on UNIXes, it throws are hefty spanner in the works when you attempt to transfer such files to a MSWin or NTFS-based storage system. (I've encountered that crap in a few git repositories already, where the other committer(s) didn't bother with NTFS/MSWin.  :-( )