Bug 705838: Improve resilience to compression bombs. (GHo note: the PDF reported there is not publicly accessible, unfortunately. So no way to doublecheck this, unless we roll our own test file ... or happen to find one in our test corpus)
When using fz_read_best to read from a stream, now allow it to take
a 'worst_case' value. If this is set, then don't trigger the
compression bomb code until we pass this value.

Accordingly, feed that value in from the image reading code with the
full size of the image, and leave it set to 0 elsewhere.

# Conflicts:
#	source/pdf/pdf-stream.c