WIP: build single-thread Win32 32-bit monolith: testing that build mode too. (DISABLE_MUTHREADS=1)
We've found over the last weeks of bulktest testing that PDFs containing JBig2 image elements are not thread-safe: those parts have already been protected by a mutex, but it begs the question: how much do we gain from executing mupdf in multithreaded mode?
At least when there's arbitrary PDFs incoming there's not a lot of gain. So the idea is this: produce cleaned-up PDFs for viewing/rendering in single-threaded mode (using multiple processes to parallelize the conversion of batches of PDF files), then in a separate binary have a multithreaded mupdf renderer which doesn't have to run into JBig2 content anymore, making that render process so much faster.

The up-front cost is a single costly conversion (compare to a single full render) but once that converted variant is cached, we can use that one to show the contents faster!

> Meanwhile, when the user would want to see the PDF stante pede, they would have to run in parallel with the cleanup process, which at least on multicore machines would be fine as neither will load the cores to the max; delaying the cleanup/conversion would potentially lead to a never-convert, so we'd better run them in parallel, then, when the user revisits the PDF at a later date do they get to see the converted/cleaned version.