further instrumenting the code while hunting down the fz_store race condition.
Now we know at least that it happens on/after the fz_free() in the `existing != NULL` branch of `fz_store_item()` but the issue is QUITE unpredictable. Working hypothesis is that another thread is able to free the `existing` instance, thus invalidating its contents, including the `val` pointer, by the time we access it.

The hard part now is finding out how it happens and what we can do about this exactly -- I fear/assume it's Artifex' code part where they inject an item "that now exists in the hash table but not yet in the list", i.e. an inserted `item` that hasn't been ref-counted up yet. (This idea would take ~3 threads cooperating to screw it up like that, *me thinks*.)