Tweak flushing of text operations in filter code.
Only output font changes etc if we are in a text object. This
avoids us eliding a BT/ET entirely and still outputting font
changes etc. This way such changes should be held over for the
next BT.

This was found during investigation for bug 705770; it does NOT
solve the bug, but tidies up the output in question.