Bug 705770: Improve redaction handling of clipping text.
If a text rendering mode with bit 2 is set, then the idea is that
the union of the text in the text object is intersected with the
existing clip path.

If we redact all the text, then we really want the clip path to be
left as being empty.

Unfortunately, the special case of the BT/ET being empty is resolved
(according to the spec) as being 'no change to the clippath', so
we need to spot this case and handle it differently.

We can't even replace the BT/ET with something like:

 0 0 0 0 re W n

as that will corrupt any path currently being generated.

So, instead, we handle the 'empty_clip_path' as being a state that
we recognise in the gstate in the filter, and simply don't forward
any calls once we reach that state.

To recognise reaching that state we have to spot a) whether we've
sent any marking characters within a BT/ET, and b) whether we've
removed any. If we reach the ET and we've removed some, and not
sent some, we know that we should now have an 'empty_clip_path'.

The nasty bit here is that we need to know if characters are
marking or not. Currently, the code assumes that space (cid 32)
is non-marking, and all others are marking.