WIP: remove performance hack for rendering pages "only roughly" when they would've taken a *lot* of time otherwise, thanks to excessive numbers of groups and/or knockout layers.
The performance logic is moved to the `fz_cookie` which is passed along, attached to the `fz_device` where available.

---

20/20 hindsight comment: this is (as was half expected) exploding in my face as the upper API layers are now getting littered by `, fz_cookie* cookie` visitor parameter additions all over: not all of the output code gets to see the `fz_device`, but we're also using the cookie to direct repair-or-fail choices on the load/parse side of things, where no `fz_device` ever roamed, and that part (pdf, xml, html, fb2, etc. parsing!) is now starting to get littered.

So we decide to take the 'slightly hackier route' I considered before: attach `cookie` to the `ctx` (context) as that will prevent function interface pollution. HOWEVER I don't like it much to have to do it this way next as that makes this kind of a hidden parameter. Anyway, this does strengthen the argument to create a separate `fz_context` for every major action: when we introduce that `cookie` to it, you'll have to **at least** refresh the cookie before you run another pdf through the same `ctx`, or you'll get cookie=progress-and-other-settings seeping in from your previous run.

Anyway, parking this off to the side...

---

PS: warning 4133 is pretty useful, so I don't know when I killed it way back when, but that one has to come back up: saves a ton in digging out the offending functions as I change their interface. Also nuked (and rightly so, I think) the function typecasts in `fz_open_reflowed_document()`. That's a keeper.