trying to get my grubby paws behind an obnoxious `abort_bgprint()` exit in the monolith `bulktest` run (!@#$%^&$%@#^$@# CENSORED).
According to MS documentation, the **only** way to hook into the `abort()` exit path is via *custom SIGABRT signal handler*, so here goes.

- https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/signal?view=msvc-170
- https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/abort?view=msvc-170

--------------

**MS** sez:

When the program is compiled in debug mode, the message box displays options to Abort, Retry, or Ignore. If the user chooses Abort, the program terminates immediately and returns an exit code of 3. If the user chooses Retry, a debugger is invoked for just-in-time debugging, if available. If the user chooses Ignore, abort continues normal processing.

^^^^^ (my feedback:) bollocks. Doesn't happen on this Win10/64 box anyway: *niente "message box" far nada zilch*. At least not when it's **important** like, e.g. when you run this tripe in the debugger.

In both retail and debug builds, abort then checks whether an abort signal handler is set. If a non-default signal handler is set, abort calls raise(SIGABRT). Use the signal function to associate an abort signal handler function with the SIGABRT signal. You can perform custom actions—for example, clean up resources or log information—and terminate the app with your own error code in the handler function. If no custom signal handler is defined, abort doesn't raise the SIGABRT signal.

^^^^^ (my feedback:) finally something good. `SIGABRT`, baby, yeah! And there's even something trying to look like a stacktrace left for one to backpedal through. :+1:

By default, in non-debug builds of desktop or console apps, abort then invokes the Windows Error Reporting Service mechanism (formerly known as Dr. Watson) to report failures to Microsoft. This behavior can be enabled or disabled by calling _set_abort_behavior and setting or masking the _CALL_REPORTFAULT flag. When the flag is set, Windows displays a message box that has text something like "A problem caused the program to stop working correctly." The user can choose to invoke a debugger with a Debug button, or choose the Close program button to terminate the app with an error code that's defined by the operating system.

^^^^^ (my feedback:) couldn't care less, now's'n me got mah SIGABRT. Ciao!