further unifying fz_eerror/fz_warn/fz_info + error/debug output channels: using the new fz_write_strings() API and the new side channel APIs to set the severity levels for each `fz_output` used by fz_error() et al, so that any default or custom backend calls can use this info to properly dispatch the error/warning/info messages.
NOTE: as `fz_write_strings()` use the printf-level mutex for the registered `fz_output` channel, you now have an in-built guarantee that each fz_error()/fz_warn/fz_info issued message is treated as an atomic message unit while it is transmitted to the output. Another benefit (side effect) of using the new `fz_write_strings()` API for these calls is that you are virtually guaranteed that, even for buffered output channels, every message will be flushed to output as each message will be terminated by a `\n`, which is detected by `fz_write_strings()` as a needs-to-flsuh signal!