Previous commit does it: render performance for some previously *slow* PDFs has improved manifold by the TK semi-revert done in the previous commit:
*semi*-revert as the code is still in there, it's just that we disabled its main feature: pushing a layer for any TK node!

Commit message of the commit that introduced this feature (see the not-merged-into-vanilla-mainline branch 'knockout' by Robin Watts, where it's part of the work done there re blending modes processing):

For now, we leave this layer-push step disabled as we have not yet found a PDF that renders incorrectly with this chunk disabled (`#if 0 ... #endif`) though admittedly we only did a very quick scan.

---

SHA-1: 2d37ec6b423c6bc7db93bee480c4910ed4bf3e6a

* Honour ExtGState TK flag.

Collecting the TK (Text Knockout) flag in the PDF interpreter is
easy. Dealing with it is somewhat harder.

The idea is that if the TK flag is set, every text object (BT/ET)
should enclose its contents in a knockout group. This both protects
separate strings in the object (e.g. (foo) TJ (bar) TJ) from
nasty effects if they overlap each other, but also individual letters
from overlapping other letters in the same string.

Accordingly, this means we need to push/pop a knockout worker group
around every single letter in each string in the worst case.

An idea that initally seems nice is to not bother pushing the knockout
group if we don't have any alpha values in play - unfortunately the
alpha values (and text rendering mode etc) can change during a BT/ET,
and hence pushing a knockout group when the first non-solid alpha
level is seen will not work.

Accordingly, the best we can do is to not push the knockout group if
there is no alpha at all on the page.