Bug 706245: Don't throw when JPXDecode filter fails, return empty pixmap.
The bug report contains a PDF where an image claims to be using a JPXDecode
filter, but the actual data is a JPEG image. OpenJPEG fails to decode the
data stream which causes MuPDF to throw an exception. That exception is not
caught anywhere, instead it reaches the application and rendering fails.

A corresponding errors in other image formats, e.g. premature end of data,
results in MuPDF rendering the page with a an empty pixmap where the image
is supposed to go.

After this commit, exceptions while decoding JPX images will also result in
empty pixmaps being returned, instead of exceptions.