qiqqa tools: adding a utility which can decode the legacy `Qiqqa.autotags` files and produce something legible (e.g. JSON) instead.
Currently we have a decoder which extracts the data correctly (tested on an old, large, Qiqqa.autotags file of mine).
Next step: use the file buffer (it's copied into RAM in its entirety) as strings buffer too; plus inject these tags and doc-hashes into a hash table of our own, which we can then use to produce the output format.