Delay parsing of JBIG2 globals until the image is loaded.
The benefit of this commit is that it will be possible to create
fz_compressed_buffer structs from within js/java since they won't
have to deal with jbig2dec specific types, but instead can create
fz_jbig2_globals with a buffer containintg the bitstream for any
jbig2 globals.

Bug688080.pdf was used as a test case to prove that rendering time
is not affected by this commit.

# Conflicts:
#	include/mupdf/fitz/filter.h
#	source/fitz/filter-jbig2.c
#	source/pdf/pdf-stream.c