fix another embarrassing bug in the XMP-and-other-streams XML parse section, where we first employ the internal parser and when that one does not deliver, wee go for the much more tolerant (or so we hope!) gumbo parser. And iff that one fails as well, we dump the offending supposedly-XML chunks to the error channel.
And that's where I b0rked: while rtrim()ming the data-to-be-dumped, the local 'rtrim' equivalent was so greedy it nuked everything but the very first line of data. Which made a bit of curious XML error log, so we investigated. And discovered this bug. Which we *then* found out in the MSVC debugger (thanks!) the stack-corruption bug fixed in the previous commit. Oops, indeed!