fix dumb error by me: forgot to check against stderr in a tesseract report function. This resulted, in the bigger `mutool_ex` app, in a **very curious** failure mode where tprintf() internals, which are sent through the fz_error() API series (mupdf, modified), invoked the windows Writefile() Win32 API ... which **failed**. Because the Win32 fixed STDERR_HANDLE was suddenly reported as the cause of WriteFile() --> error 6 (Invalid Handle)!
Now it gets dirtier...

When this happens, code from previous run-ins with serious bad shite would try to report this error to the fz_stddbg() 'standard debug' channel (which is a thing that does exist on MSWin and *is* used by your truly, **plus** the same error report would go to fz_stderr.

**whoops!**

That one is *invalid* by now, according to the OS, (thanks to an inadvertent `fclose(f)` where `f == stderr`) so we got into a cycle which uncovered another bug as this failure mode was yet unexpected: we were in a fz_try/fz_catch C-style exception frame, BUT we were also inside *another* C-style exception being thrown, so the cycle-detection logic, which was engineered to prevent a stack overrun due to crazy galloping scenarios like this, would simply attempt to report the *now fatal & new* error to stderr, where it ... fails.

A side-effect of the bug(s) in that code resulted in the fz_throw() to **return to the caller**, which was very much *not* expected and resulted in an infinite loop around the attempt to get WriteFile() do write some text, WriteFile() telling us it failed, length written: `n == 0`, loop says: keep trying while `n > 0` and there you are.

That loop bit was previously devised to cope with another nasty mishap we had a very tough time debugging, where it turned out that when you use stdout/stderr echo-ing applications via any type of syscall which 'redirects' those stdout+stderr handles to application handles, i.e. Win32 *pipes*, so the calling application can get its grubby paws on your invoked executable's stdout/stderr blatherings, things go very bad indeed when the **syscall-invoking application** happens to be (relatively) slow with the stdout/stderr pipe redirection emptying: as that particular app (Qiqqa up to v83) didn't ever have async pipe stream reading implemented before, it thus turned out that MS Win64 has somewhere around 64K pipe channel buffer for your stdout/stderr redirects via Win32 pipes and WriteFile, which is always at the bottom/core of any printf()/fprintf(stderr), will then **hang**. At the most curious of times and places.
Unless... you 'improve' your stdout/stderr writing mechanism with some code that is meant to cope with 'force majeure' failures like this, i.e. using a WriteFile that's time-bound, i.e. will now **not hang** any more but instead report some failure diagnostic (with the help of a few friendly Win32 APIs) about *why* it didn't succeed in delivering the goods.

It all works great, did great under various adversarial circumstances, so case closed.

Until today, where we didn't realize we just fclose()d stderr and got another type of error. Now I knew I would *ideally* have to be pretty pedantic in the failure-mode code lines around the fitz/mupdf C-style exception handler, but previously didn't think that level of detail was necessary (it wasn't) and when done, it would look butt-ugly in the code (it does) as every 'independent' line/chunk that is at risk must be wrapped in its own try/catch while we travel down the rabbit hole of failure-upon-failure as failures are want to be reported... which, of course, when you fclose(stderr) and the OS takes you at your word, is cause for a lot of pain.

So the fix is in, the fz_try/fz_catch + error/debug channels now support this scenario and 'survive': ultimately they'll kick in the debugger or call exit(666) in a final attempt to be nice about it.

Meanwhile, the tesseract code gets the other fix: fclose prevention for stderr.

It is, however, part of a bigger fix elsewhere...