Change forward page map to be a list of object references.
Currently, the forward page map is a list of object numbers.
This means that when we pdf_lookup_page_obj, we resolve the
object to get a borrowed ref (the xref reference) to the
required object.

Unfortunately, a lot of code that calls this assumes that
we'll get a borrowed reference to an indirection object that
points to the page object (i.e. they want a page ref, not the
page object).

To fix that, we move the array to be object references (as
loaded from the Page tree, so indirection objects) rather than
object numbers.

This should fix the problem where PyMuPDF's xref member for
the Page object was always returning 0.

# Conflicts:
#	include/mupdf/pdf/document.h
#	source/pdf/pdf-page.c