RFC: Attempt to not crash when manipulating unbound pdf objects.
This is not a full-ended solution since indirect objects, arrays
and dictionaries all hold references to pdf_document *, but when
we drop that document, we don't know that this happened.

Is it ok for pdf objects to hold references to the document such
that it cannot be dropped under the feet of the pdf object?
Probably not. I think we've tried that and failed, but I don't
remember why.