Bug 699212: Prefer default method/key length based on crypt ver/rev.
Previously if crypt filters' method/key lengths were inconsistent
with the expectations of the crypt version/revision MuPDF would prefer
those given by the crypt filter.

The file in the bug specifies version 5 revision 6 encryption.
According to spec revision 6 implies version 5, and version 5
is always a 256 bit key with the AES cryption method. The crypt
filters in the file specify AESV2. But according to PDF 1.7
extension level 3 AESV2 is only valid for version 4 revision 4
encryption!

Since MuPDF preferred the crypt filters' encryption method AESV2
and key length 128 bits, an exception was triggered later when
the key length was found not to be 256 bits.

After this commit, the keylength and method implied by ver/rev in
the crypt dictionary will be assumed.

# Conflicts:
#	source/pdf/pdf-crypt.c