Bug 706403: Workaround setjmp problems on some windows builds.
SumatraPDF reports to us that they are consistently seeing
problems with setjmp/longjmp on 64bit windows builds.

Scouring the incomplete reports on this, it seems like there
may be a problem to do with setjmp behaving differently on
DLL and non-DLL built code. It is therefore my working
hypothesis that there may be a problem when setjmp/longjmp
is used over a DLL boundary.

Adopt the proposed fix, because it seems harmless.

# Conflicts:
#	source/fitz/error.c