Makefile Makerules: Allow builds for Pyodide/wasm.
In Pyoide builds we use new `scripts/bin2obj.py` instead of `ld -b binary`
because the latter is not supported by pyodide's replacement `ld`.

In Makerules, cope with HAVE_LIBCRYPTO set in environment. If HAVE_LIBCRYPTO is
set in environment (instead of the make command itself), make will overwrite it
with `=` or `:=`. So we need to explicitly look at whether $(HAVE_LIBCRYPTO) is
empty before setting with $(shell pkg-config ...).