Update default sRGB and sGray ICC profiles.
The default Gray ICC profile was using a simple gamma 1.8 TRC.
The gray sGray profile is supposed to match the sRGB TRCs with the
same color primaries.

Update both the gray and RGB to more compact profiles with the same or
better accuracy in the TRC curves.

https://github.com/saucecontrol/Compact-ICC-Profiles

These are the v2-magic profiles.