Bug 706762: OSS Fuzz 69497: Fix SEGV after pdf repair.
Suppose we have to read object 'x' that is supposed to be in
object stream 'y'. While reading 'y', we rewrite the type of
'x' in the xref table to avoid recursion. We then put this
back after the read of 'y' completes.

Unfortunately, if 'y' is corrupt, we can trigger a file
repair. During the course of the repair, we can discover that
'x' doesn't actually exist at all, resulting in an xref table
that no longer contains an entry for it.

When we exit from reading 'y', the existing code tries to
put the type of 'x' back, causing a crash. The fix is simply
to make sure that 'x' has an entry, and has been rewritten
first.