Add support for CropBox, TrimBox, BleedBox, and ArtBox.
Clip page contents to CropBox if it is smaller than the MediaBox.

Add a new function fz_bound_page_box to request bounds for using a box other
than the default MediaBox.

Add a "box" argument to PDF-specific pdf_new_pixmap_from... helper functions.

The "fitz" coordinate space is unchanged; it is still left-handed with the
origin based on the MediaBox.

# Conflicts:
#	include/mupdf/fitz/document.h
#	platform/gl/gl-main.c
#	platform/wasm/lib/mupdf.c
#	platform/wasm/lib/mupdf.js
#	source/fitz/document.c
#	source/pdf/pdf-run.c
#	source/tools/mudraw.c