Bug 707079: Do not scale color key for indexed images.
The PDF in the bug has an image which has 4 bits per pixel and
is indexed. So the indices are in the range 0..15.

While samples for normal images are scaled to 8 bits per pixel.
Thus the color key must be correspondingly scaled to the 8 bit
interval so it can be used to compare each sample.

The indices for indexed images are not scale, they are copied
verbatim into each sample byte in the pixmap. Therefore the
color key need not be scaled to the 8 bit interval when it is
compared to each sample in the pixmap.

This commit simply propagates an existing flag indicating whether
the image is indexed and skips scaling of the color key for
indexed images.