Bug 707199: mutool pages 'infinite' loop.
The proof of concept here claims to have millions of pages, but
only actually has one. We were failing to spot this when
loading the page tree, and then pratfalling later on.

Improve the page tree loading to spot the real number of pages
and to fix the file.

Also, improve mutool pages so that it doesn't try to infinitely
loop across the millions of non-existent pages.

# Conflicts:
#	source/pdf/pdf-page.c