Bug 706995: Fix regression in spot colour handling.
When painting a pixmap, we can cope with painting a pixmap with
no spots onto a destination with spots, but if they both have
spots, the spots must match.

Previously our test was too strict, and we were forcing pixmaps
with no spots to be converted before being painted.

Also, fix a problem whereby our special case 'gray to rgb' plotters
were being triggered when we were going from 'gray to gray + 2 spots'.