updating MSVC project files; moving towards another way to remove "failed projects" (repositories of interest that are deemed "too much work to get working in a way that I'ld trust"(tm) from the owemdjee set as that one is slowly exploding without a well-behaved cleanup phase.
The trouble is: I'm still not happy about the cleanup phase processes I've been using / trying to date; this is another attempt to try/find a process of administration that works satisfactory for me and the way my brain / mental processes work; after all, let's face it, I'm the main developer working at the project(s) which wish to to use these repositories in some R&D mix and the process is still ... unsatisfactory. So the search for a better suited workflow continues. I've tried a few ways others have shown to cope with this or a very similar challenge and those ways simply do not work for me, at all. The really sad part is I'm still not able to properly, *precisely*, formulate/verbalize the problem+goal, which makes finding the workflow (= solution) rather ... haphazard. Soit. This is where I am, this is what we've got and I still need to move forward, so that ends this lament and on to the next trial&error: the big list (a.k.a. cummulative list of source trees of interest ~= "repositories") is *copied* to two target lists: the "may matter" list and the (important to me!) "failed attempts/ideas" list: the former SHOULD be filtered to only include projects that I deem *potentially viable* (easy enough to get working in monolithic build mode; some other vague criteria, etc.) and the "failed ideas" one SHOULD be the *complement*, i.e. listing all the repo/projects that still may be of interest, but are deemed "too much fuss to bother with any further right now". One might call it the "cut your losses" list. or maintenance of the lists: the m_dev_list master cummulative set is leading; the "may matter" list is he copy thereof where any new and old material gets filtered by throwing it out of the SLN (solution); the "failed ideas" list is the *exact* complement, hence can be almost-auto-produced by taking the dev_list master, copy it to "failed ideas" and *subtracting* the "may matter" subset next. Which will give us the "failed ideas" set.

What SHOULD be added to the "failed ideas" solution (or the owemdjee repo itself) is a set of notes describing what I've been doing, what I saw/experienced/feel about the repo XYZ, which are the grounds WHY I ditched the project and moved it into the "failed ideas" realm. (Of course I could go and pick it up later; a fundamental issue I have with all the approaches to date is that they do not cope well with adaptiveness in both directions: throwing out/away is relatively easy, but moving it back in / picking up a left-off is far from easy; at least harder than throwing out. I don't mind failures staying around because those "failures" may very well be "failures at that historic time" and ecome "good stuff" later on; the inverse progression can occur as well (note to self: recall my old OpenSSL overlay work (not open source) that has gone the way of the dodo.) This bidirectional movement in my project lifetimes is something I haven't seen elsewhere; nobody I've spoken to over the decades identified it as a pressing problem, while I keep running into it; it's only very recently that I've become slightly better at verbalizing/identifying this as the fundamental concept that is the source of my irritation at existing workflows/processes.

End of lament.

Written at night so reckon with a odd lucity/sensible/nonsensible mix in the paragraphs above. Not optimal, but now is the time and this should be written before I sleep, for it should persist, even when present in b0rked form only.