Tweak mutool clean's handling of images.
When cleaning with flags such as -diz we used to leave images alone.

This means that JPEGs stay as JPEGs (what we want, generally), but
it also means that flated images remain flated - not what we want
when we are experimenting with better compression methods.

Now, we keep images that are compressed with an image codec (Fax,
JPEG, JBIG2 etc), and recompress other ones.

# Conflicts:
#	source/pdf/pdf-write.c