Bug 707455: Parse CSS case-insensitively.
The bug is mainly about a crash, but mentions that browsers parse the
CSS string "!important" case-insensitively. According the to
 https://www.w3.org/TR/2007/CR-CSS21-20070719/syndata.html#characters
style sheets are case-insensitive in general, with exceptions for ids
and classes and font names.

This commit uses the newly introduced fz_strcasestr() to replace
uses of strstr() in the css parser and replaces strcmp() with the
corresponding fz_strcasecmp().