Add support for Rich Content in FreeText annotations.
Uses the Story mechanism to place the content within the rectangle.

This support is only used if FZ_ENABLE_HTML_ENGINE is defined
on building. i.e. by default, if you disable all the HTML using
document agents, then we will not support Rich Content to avoid
pulling in the entire HTML agent just for this.