Bug 707816: Correct undo/redo issues.
When we alter object n, we check for it being in the 'pending'
list already, so that we don't record 2 different changes for
the same object.

Unfortunately, when we have nested operations, there can be
several 'pending' journal entries. We should be checking in the
last one (the pending_tail), not the first one!