Bug 707824: Fix SVG generation of files with layers.
We send layers into the SVG as <g id="layer%d" ...> tags.

Unfortunately, this plays badly with clip paths, which are also sent
as groups. Because layers and clips don't necessarily nest nicely in
PDF, we can get problems.

Here, we revise the layer handling in SVGs. Firstly, we keep the
layer details we've seen before in a dict. This enables us to
send the same layer number for every instance of the same layer
name, rather than the incorrect number we were sending before.

Then we keep the current layer nesting in a stack, and remove that
before changing clip paths, replacing it afterwards.