this was us debugging a very obnoxious segfault + heap corruption. The count check in the output writer turns off all console output, thus speeding up the test run enormously (lots of debug output traveling through during this test scenario - the debug output from tesseract triggered the issue, so the quickest way to shut up and still have a chance of catching the bugger was turning off the console write delay. It worked.
Also has other quick/panic hacks while we were hunting a segfault which occurred *randomly* for the *same test*(!) -- hence thread (non-)safety was suspected at first, but after killing anything that could become a thread, the error persisted and the next approach was to suspect the Win32 random address anti-hacker feature: we COULD get random behaviour if the error is some sort of memory corruption that depends on the address where the app is loaded. That turned out to be the case as also turning on the MSVC crtdbg features to near-full throttle, the segfault became preceded by a much more predictable heap corruption report about the top of a free chunk being nuked. The rest is in tesseract: my mistake with the `Image` class NOT being a Pix pointer owner. (Fun bit: it turns out that offending assignment operator got invoked exactly *once* during the entire app run, and that's when the heap got corrupted through unsuspected double-free action.   *Ouch!*